Storing the quantity simulation flag in the SIMDD
In the SIMDD file, there is a parameter called 'quantity_type' which is used to specify which Quantity subclass instance is to be used to generate simulation data for the quantity/attribute. This is all handled in the setup_sim_quantities method.
The tests for the simdd parser were modified to accomodate these new changes.