Raising the ValueError exception for missing parameters
Raising a 'ValueError' exception instead of a 'NotImplemented' exception when we encounter an attribute with a data simulation quantity_type not 'ConstantQuantity' having no mininum and/or maximum value specified.
Included a new attribute for the parser class, so that we can store the name of the configuration file that is parsed to make it easier to debug.

JIRA: LMC-210