LMC-219 Testing the Improvement of the tango sim generator conform with data driven simulators:
The tango sim generator generates a python.py server file with the same name as the server name which contains
all the neccesessary info to run a tango device server simulator. In the tests, a simulator is generated and
we test if it actually matches the specification provided in the simulator description file.

- A temporay directory is created where the generated tango simulator executable is store

- A temporary tango database file is also created and stored in the same temp directory which at the end of the
test is removed

- The tango device server simulator with its corresponding simulator test device are then run in a separate process on tcp port 12345

- A client is then instatiated cnnecting to the running device server, the testing of the device is perfomed

- Finally the temp directory is the remove and the device server process is terminated

Note:
At the moment an xmi file is a requirement since it contains all attribute infomation and most importantly tango class name.