Addressing PR#6 review comments
Refactored the (pan, tilt, zoom, focus)_position action handlers to get rid of the nested try/except statements and improve code readability.
Converting the (pan, tilt, zoom, focus)_postion command argument values to the same data type as their correspond quantity/attribute data types (which are floats, not ints).
Refactored the preset_(set, goto, clear) action handlers, so that the device simulator can still take a snapshot of the values of the existing position quantity's (pan, tilt and/or zoom) and be able to revert back to the preset state without any problems. Position quantities not in the model will be logged.
Created a simple exception class for each of the override classes.
Raising exceptions for invalid optional command arguments.
Raising exceptions for commands that need to access but cannot find a model quantity during their execution.
Updated the docstrings for the action handler functions/methods.
Added a functionality in the 'camerapoweron' action handler that changes the value of the TANGO device 'State' attribute to the argument value passed the 'CameraPowerOn' command.
Modified the TANGO device class 'read_attribute' method to update the device attribute values only when the device state in not 'OFF'.