Adding tests source simulator info
override_class.py: Passing a second positional argument to the Quantity set_val method in order avoid getting a TypeError.
sim_xmi_parser.py: Added the property_group parameter to the get_reformatted_properties_metadata method signature (also in the simdd_json_parser.py file). Redefined the get_reformatted_properties_metadata method to return either the device or class properties dictionary. When generating a TANGO device simulator using both the XMI and SIMDD files, it so happens that when populating the model quantities using the information from both the parsed files, the quantities' meta data gets overwritten instead of being updated. So to work around that I have modified the setup_sim_quantities method to check if a quantity already exists in the model sim_quantities and if it does, it takes its meta data and updates it accordingly.
test_simdd_json_parser.py: Wrote tests that check for data consistencies but not including duplication between the XMI and the SIMDD file.