Updating the model quantity properties
In the case of using the xmi and simdd config files, the xmi information is used to initialize the quantity meta and then updated by that of simdd. Before the model attribute props dict is updated, the parameter keys with no values specified from the attribute props template are removed.
i.e  All optional parameters not provided in the SIMDD