Allow custom logger to be passed to model
Different projects often have different logging requirements.
This change allows a logger instance to be passed to the
models.  This will allow it to be used in the override
class, if necessary.

The logger instance is not passed directly to the override
class.  This is because that would result in a change to the
API that is unnecessary.  All calls to the override class
include the `model` instance as a parameter, so the `logger`
instance can be accessed from the `model`.

Some of the examples have been updated to reflect this
new usage.

JIRA (SKA):  SAR-155