SAR-166 Configure all the attribute's default properties (#133)
* SAR-166 Configure all the attribute's default properties.
This enables us to have a running TANGO device with all the attributes' properties configured as is specified in the interface configuration files (.fgo, SimDD and/or .xmi).

* SAR-166 Added default values for some attribute properties.
This is to make the generator fail safe when some of the non-required attribute properties are not specified in the interface configuration files.

* SAR-166 Set default property values to empty strings.
The documentation seems to be inaccurate https://pytango.readthedocs.io/en/stable/server_api/server.html?highlight=attribute#tango.server.attribute

* SAR-166 Set 'maxValue' and 'minValue' default value to empty strings
These are supposed to be empty for an attribute with DevEnum data type.

* SAR-166 Added a comment for not setting enum_labels.
Setting the 'enum_labels' attribute results in the error: Python argument types in
    None.None(UserDefaultAttrProp, list)
did not match C++ signature: None(Tango::UserDefaultAttrProp {lvalue}, std::__cx^C1::basic_string<char, std::char_traits<char>, std::allocator<char> >)

* SAR-166 Resolved a TODO.
Changed 'polling' to 'period' on the device_attribute dict generated by the fandango file parser to keep the same interface across file parsers.

* SAR-166 Added more info to error messages.

* SAR-166 Convert 'period' to a string type.
The 'UserDefaultAttrProp' expects it to be a string.

* SAR-166 Added error handling when setting attribute properties.
This will ensure that the 'initialize_dynamic_attributes' method doesn't ternminate when there is an exeception raised and continues to add subsequent attributes. Also set the value for 'event_properties' using its setter method.

* SAR-166 Changed logging level to debug.
This should remove all the noise in the logs when running the device simulator.