Updated the class instance prefix to be more definitive
When the startswith method is called with the  prefix 'Sim' it resulted to true for 'SimControl'. This results in the sim_actions
not getting overriden with the correct action_handler as it is not defined in the provided 'SimControl' class.
This is why only one of the devices' commands where getting overriden.

Signed-off-by: Katleho Madisa <katleho.madisa47@gmail.com>