#153 Fixed SPECTRUM attr creation for SimDD file
This converts the attribute's data_format property to the tango.AttrDataFormat enum instead of using the string value specified in the SimDD file. The attribute was getting added as a SCALAR type because of that.
I have also specified some default values to use when creating an instance of attribute in case the user doesn't specify them in their SimDD file. This will prevent the tango device server from crashing when starting up.
A test is also added to confirm that the library supports creation of SPECTRUM type attributes from the SimDD file.