#153 Fixed bug reading SPECTRUM attribute
Refactored the model.py module. This is to ensure that the default value or value added to the quantity is according to the data_type and data_format properties specified in the configuration data files.
Also added a test to confirm that we can read the SPECTRUM attribute on the device if it is specified in the SimDD file.