#153 Fixed SPECTRUM attribute creation for SimDD file (#154)
* #153 Fixed SPECTRUM attr creation for SimDD file

This converts the attribute's data_format property to the tango.AttrDataFormat enum instead of using the string value specified in the SimDD file. The attribute was getting added as a SCALAR type because of that.
I have also specified some default values to use when creating an instance of attribute in case the user doesn't specify them in their SimDD file. This will prevent the tango device server from crashing when starting up.
A test is also added to confirm that the library supports creation of SPECTRUM type attributes from the SimDD file.

* #153 Standardized file parser data

I have changed the attribute property 'initial_value' to 'value'. This ensures that the data from all file parsers is the same and makes it easy to process without making special provision for each parser.

* #153 Dropped support for the XML config file

The file format is no longer used anymore and was added for experimental purposes.

* #153 Fixed bug reading SPECTRUM attribute

Refactored the model.py module. This is to ensure that the default value or value added to the quantity is according to the data_type and data_format properties specified in the configuration data files.
Also added a test to confirm that we can read the SPECTRUM attribute on the device if it is specified in the SimDD file.

* #153 Added more tests for the SPECTRUM attributes

Testing Spectrum attributes of different data types, String and Boolean, respectively.