#153 Support IMAGE attribute addition (#155)
* #153 Enabled creation of IMAGE attribute
This change just enables the creation of attributes with a data format IMAGE. This is added as a static attribute before the device is started.

* #153 Updated and extended the tests
Added tests to check that the attributes with a data format IMAGE/SPECTRUM which are specified in the configuration files are added correctly on the device and that they are readable.

* #153 Unsorted the enum labels
Removed the sorting of enum labels. This is to preserve the order that was specified in the xmi configuration file.

* #153 Updated tests
This is to ensure that the order of the enum labels is preserved.