Added support for int types in the SimDD config
These updates the SimDD.schema to support more types, especially int types. Note that DevInt is no longer a supported type.
The rest of the changes are to test that we are able to generate a device and read the int type attributes. I have also
updated the tango_sim_generator module to ensure that the  attribute write type defaults to READ if it is not specified
in the config file.

Resolves: #156

Signed-off-by: Katleho Madisa <katleho.madisa47@gmail.com>